% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatos_crescimento.R
\encoding{UTF-8}
\name{ajustar_curva}
\alias{ajustar_curva}
\title{Ajuste canônico de modelos de regressão não-linear}
\usage{
ajustar_curva(
  dados,
  var_y,
  var_x,
  tipo = c("potencia", "von_bertalanffy", "logistico", "exponencial", "polinomial",
    "logaritmica")
)
}
\arguments{
\item{dados}{data.frame/tibble com as variáveis.}

\item{var_y}{nome da variável dependente (string).}

\item{var_x}{nome da variável independente (string).}

\item{tipo}{tipo do modelo a ser ajustado (padrão "potencia").}
}
\value{
Uma lista estruturada com o modelo ajustado (nls), coeficientes,
  métricas de ajuste (pseudo-R², RSE, AIC) e validação de normalidade dos resíduos.
}
\description{
Ajusta um modelo não-linear (potência, Von Bertalanffy, logístico, exponencial,
polinomial quadrático ou logarítmico) a partir de um conjunto de dados.
}
