% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatos_crescimento.R
\encoding{UTF-8}
\name{curva_predita}
\alias{curva_predita}
\title{Grade fina de pontos preditos para gráficos}
\usage{
curva_predita(r, n = 200)
}
\arguments{
\item{r}{lista retornada por \code{ajustar_curva}.}

\item{n}{número de pontos na grade (padrão 200).}
}
\value{
Um data.frame contendo as colunas de X e Y com os valores estimados da curva.
}
\description{
Grade fina de pontos preditos para gráficos
}
