% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_amostragem_tabelas.R
\docType{data}
\name{dicionario_variaveis_amostragem}
\alias{dicionario_variaveis_amostragem}
\title{Dicionario de variaveis dos conjuntos de amostragem}
\format{
Um data frame (tibble) com 6 colunas:
\describe{
  \item{base}{Nome do conjunto de dados ao qual a variavel pertence.}
  \item{variavel}{Nome da variavel, em portugues.}
  \item{descricao}{Descricao da variavel.}
  \item{unidade}{Unidade de medida, quando aplicavel (NA caso contrario).}
  \item{tipo}{Tipo da variavel (inteiro, numerica, categorica, logica, etc.).}
  \item{uso_na_amostragem}{Papel na amostragem (estrato, variavel de interesse, etc.).}
}
}
\source{
Elaboracao propria a partir das documentacoes dos pacotes FSAdata e MASS.
}
\usage{
data(dicionario_variaveis_amostragem)
}
\description{
Tabela que descreve, em portugues, cada variavel dos conjuntos
\code{walleye_erie}, \code{snapper_hg} e \code{crabs_morfometria}: nome,
descricao, unidade, tipo e papel na amostragem (variavel de interesse ou
estrato). Util como apoio didatico em amostragem aleatoria simples (AAS) e
amostragem estratificada proporcional.
}
\examples{
data(dicionario_variaveis_amostragem)
head(dicionario_variaveis_amostragem)
subset(dicionario_variaveis_amostragem, base == "walleye_erie")
}
\keyword{amostragem}
\keyword{datasets}
\keyword{dicionario}
