% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatos_crescimento.R
\encoding{UTF-8}
\name{equacao_curva}
\alias{equacao_curva}
\title{Equação formatada do modelo ajustado}
\usage{
equacao_curva(r)
}
\arguments{
\item{r}{lista retornada por \code{ajustar_curva}.}
}
\value{
Uma string com a equação matemática estimada em português.
}
\description{
Equação formatada do modelo ajustado
}
