% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatos_anova.R
\name{mostrar_anova}
\alias{mostrar_anova}
\title{Tabela de ANOVA formatada}
\usage{
mostrar_anova(r)
}
\arguments{
\item{r}{lista devolvida por \code{calcular_anova}.}
}
\value{
Um tibble com a tabela ANOVA (fonte, gl, SQ, QM, F, p).
}
\description{
Tabela de ANOVA formatada
}
