% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatos_crescimento.R
\encoding{UTF-8}
\name{mostrar_coefs_curva}
\alias{mostrar_coefs_curva}
\title{Tabela de coeficientes do modelo ajustado}
\usage{
mostrar_coefs_curva(r)
}
\arguments{
\item{r}{lista retornada por \code{ajustar_curva}.}
}
\value{
Um tibble com os parâmetros estimadores da curva.
}
\description{
Tabela de coeficientes do modelo ajustado
}
