% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatos_crescimento.R
\encoding{UTF-8}
\name{mostrar_metricas_curva}
\alias{mostrar_metricas_curva}
\title{Tabela de métricas globais de ajuste}
\usage{
mostrar_metricas_curva(r)
}
\arguments{
\item{r}{lista retornada por \code{ajustar_curva}.}
}
\value{
Um tibble com as estatísticas descritivas de qualidade global do ajuste.
}
\description{
Tabela de métricas globais de ajuste
}
