% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatos_crescimento.R
\encoding{UTF-8}
\name{mostrar_normalidade_curva}
\alias{mostrar_normalidade_curva}
\title{Tabela de normalidade dos resíduos (Shapiro-Wilk)}
\usage{
mostrar_normalidade_curva(r)
}
\arguments{
\item{r}{lista retornada por \code{ajustar_curva}.}
}
\value{
Um tibble de uma linha com o teste de Shapiro-Wilk.
}
\description{
Tabela de normalidade dos resíduos (Shapiro-Wilk)
}
