% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatos_anova.R
\name{mostrar_pressupostos}
\alias{mostrar_pressupostos}
\title{Tabela de validacao dos pressupostos da ANOVA}
\usage{
mostrar_pressupostos(r)
}
\arguments{
\item{r}{lista devolvida por \code{calcular_anova}.}
}
\value{
Um tibble com os testes de Shapiro-Wilk e Bartlett.
}
\description{
Tabela de validacao dos pressupostos da ANOVA
}
