% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatos_regressao.R
\name{mostrar_pressupostos_regressao}
\alias{mostrar_pressupostos_regressao}
\title{Tabela de pressupostos da regressão linear}
\usage{
mostrar_pressupostos_regressao(r)
}
\arguments{
\item{r}{lista retornada por \code{calcular_regressao} ou modelo lm.}
}
\value{
Um tibble de duas linhas contendo os testes de pressupostos.
}
\description{
Retorna uma tabela estruturada (tibble) com a verificação de normalidade de resíduos
e homocedasticidade de variâncias.
}
