% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatos_anova.R
\name{mostrar_tukey}
\alias{mostrar_tukey}
\title{Tabela de comparacoes multiplas de Tukey HSD}
\usage{
mostrar_tukey(r)
}
\arguments{
\item{r}{lista devolvida por \code{calcular_anova}.}
}
\value{
Um tibble com as comparacoes par a par de Tukey HSD.
}
\description{
Tabela de comparacoes multiplas de Tukey HSD
}
