% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatos_crescimento.R
\encoding{UTF-8}
\name{relatar_curva}
\alias{relatar_curva}
\title{Relato textual automatizado em português}
\usage{
relatar_curva(r, label_y = NULL, label_x = NULL)
}
\arguments{
\item{r}{lista retornada por \code{ajustar_curva}.}

\item{label_y}{descrição da variável dependente (opcional).}

\item{label_x}{descrição da variável independente (opcional).}
}
\value{
Uma string com o relato estatístico interpretativo formatado em Markdown.
}
\description{
Relato textual automatizado em português
}
