% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatos_regressao.R
\name{relatar_regressao}
\alias{relatar_regressao}
\title{Relato textual automatizado da regressão simples}
\usage{
relatar_regressao(r, label_y = NULL, label_x = NULL)
}
\arguments{
\item{r}{lista retornada por \code{calcular_regressao} ou modelo lm.}

\item{label_y}{rótulo de texto descritivo para Y (opcional).}

\item{label_x}{rótulo de texto descritivo para X (opcional).}
}
\value{
Uma string de texto em português.
}
\description{
Gera um texto descritivo científico em português a partir dos resultados ajustados da regressão.
Suporta compatibilidade retroativa.
}
