% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatos_teste_t.R
\name{relatar_teste_t_pareado}
\alias{relatar_teste_t_pareado}
\title{Relato do teste t pareado}
\usage{
relatar_teste_t_pareado(
  x1,
  x2,
  nome1 = "Antes",
  nome2 = "Depois",
  conf = 0.95,
  label_y = "a variável de interesse"
)
}
\arguments{
\item{x1}{vetor numérico (momento 1 / antes).}

\item{x2}{vetor numérico (momento 2 / depois).}

\item{nome1}{rótulo do momento 1.}

\item{nome2}{rótulo do momento 2.}

\item{conf}{nível de confiança do IC (padrão 0,95).}

\item{label_y}{rótulo da variável de interesse.}
}
\value{
Uma string com marcação Markdown para itálico no docx.
}
\description{
Relato do teste t pareado
}
