% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatos_crescimento.R
\encoding{UTF-8}
\name{tipo_curva_label}
\alias{tipo_curva_label}
\title{Rótulo amigável do tipo de curva}
\usage{
tipo_curva_label(tipo)
}
\arguments{
\item{tipo}{string identificadora do tipo de curva ("potencia", "von_bertalanffy", "logistico", "exponencial", "polinomial", "logaritmica").}
}
\value{
Uma string com o nome formatado do modelo em português.
}
\description{
Rótulo amigável do tipo de curva
}
