% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_amostragem_reais.R
\docType{data}
\name{walleye_erie}
\alias{walleye_erie}
\title{Dados biológicos de Walleye do Lago Erie}
\format{
Um data frame com variáveis em português:
\describe{
  \item{id_peixe}{Identificador sequencial do peixe.}
  \item{id_lance_rede}{Identificador do lance ou conjunto de rede.}
  \item{codigo_local}{Código original do local regional.}
  \item{local_regional}{Descrição regional do local de coleta.}
  \item{grade_amostral}{Grade amostral de coleta.}
  \item{ano}{Ano de coleta.}
  \item{comprimento_total_mm}{Comprimento total, em milímetros.}
  \item{peso_g}{Peso corporal, em gramas.}
  \item{sexo}{Sexo do peixe.}
  \item{maturidade}{Condição de maturidade.}
  \item{idade_anos}{Idade estimada, em anos.}
  \item{classe_comprimento}{Classe de comprimento criada por quartis.}
}
}
\source{
FSAdata::WalleyeErie2. Título original: "Biological data for Walleye from Lake Erie, 2003-2014."
}
\usage{
walleye_erie
}
\description{
Conjunto de dados real com informações biológicas de Walleye
(*Sander vitreus*) coletados no Lago Erie entre 2003 e 2014.
A base foi preparada para fins didáticos em amostragem aleatória
simples e amostragem estratificada proporcional.
}
\keyword{datasets}
