% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatos_teste_t.R
\name{calcular_teste_t}
\alias{calcular_teste_t}
\title{Calculo canonico do teste t para uma amostra}
\usage{
calcular_teste_t(x, mu = 0, conf = 0.95)
}
\arguments{
\item{x}{vetor numerico.}

\item{mu}{valor de referencia (media hipotetica sob H0).}

\item{conf}{nivel de confianca do IC (padrao 0,95).}
}
\value{
Uma lista com as estatisticas do teste.
}
\description{
Calculo canonico do teste t para uma amostra
}
