% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatos_teste_t.R
\name{calcular_teste_t_ind}
\alias{calcular_teste_t_ind}
\title{Calculo canonico do teste t para duas amostras independentes}
\usage{
calcular_teste_t_ind(formula_obj, data, equal_var = FALSE, conf = 0.95)
}
\arguments{
\item{formula_obj}{formula y ~ x.}

\item{data}{data.frame.}

\item{equal_var}{logico; FALSE usa Welch.}

\item{conf}{nivel de confianca do IC.}
}
\value{
Uma lista com as estatisticas.
}
\description{
Calculo canonico do teste t para duas amostras independentes
}
