% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatos_teste_t.R
\name{calcular_teste_t_pareado}
\alias{calcular_teste_t_pareado}
\title{Calculo canonico do teste t pareado}
\usage{
calcular_teste_t_pareado(x1, x2, conf = 0.95)
}
\arguments{
\item{x1}{vetor numerico (momento 1 / antes).}

\item{x2}{vetor numerico (momento 2 / depois).}

\item{conf}{nivel de confianca do IC (padrao 0,95).}
}
\value{
Uma lista com as estatisticas do teste.
}
\description{
Calculo canonico do teste t pareado
}
