% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatos_teste_t.R
\name{calcular_teste_t_pareado}
\alias{calcular_teste_t_pareado}
\title{Cálculo canônico do teste t pareado}
\usage{
calcular_teste_t_pareado(x1, x2, conf = 0.95)
}
\arguments{
\item{x1}{vetor numérico (momento 1 / antes).}

\item{x2}{vetor numérico (momento 2 / depois).}

\item{conf}{nível de confiança do IC (padrão 0,95).}
}
\value{
Uma lista com as estatísticas do teste.
}
\description{
Cálculo canônico do teste t pareado
}
