% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatos_teste_t.R
\name{flextable_ocean}
\alias{flextable_ocean}
\title{Formata um tibble na identidade visual Ocean Gradient (flextable, docx)}
\usage{
flextable_ocean(tab)
}
\arguments{
\item{tab}{um data.frame/tibble.}
}
\value{
Um objeto flextable formatado.
}
\description{
Formata um tibble na identidade visual Ocean Gradient (flextable, docx)
}
