% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatos_teste_t.R
\name{mostrar_teste_t}
\alias{mostrar_teste_t}
\title{Tabela de resultados do teste t para uma amostra}
\usage{
mostrar_teste_t(x, mu = 0, nome = "Variável", conf = 0.95)
}
\arguments{
\item{x}{vetor numérico.}

\item{mu}{valor de referência (média hipotética sob H0).}

\item{nome}{rótulo da variável.}

\item{conf}{nível de confiança do IC (padrão 0,95).}
}
\value{
Um tibble de uma linha com os resultados.
}
\description{
Tabela de resultados do teste t para uma amostra
}
