% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatos_teste_t.R
\name{relatar_teste_t_pareado}
\alias{relatar_teste_t_pareado}
\title{Relato do teste t pareado}
\usage{
relatar_teste_t_pareado(
  x1,
  x2,
  nome1 = "Antes",
  nome2 = "Depois",
  conf = 0.95,
  label_y = "a variável de interesse"
)
}
\arguments{
\item{x1}{vetor numerico (antes).}

\item{x2}{vetor numerico (depois).}

\item{nome1}{rotulo do momento 1.}

\item{nome2}{rotulo do momento 2.}

\item{conf}{nivel de confianca do IC.}

\item{label_y}{rotulo da variavel de interesse.}
}
\value{
Uma string com marcacao Markdown.
}
\description{
Relato do teste t pareado
}
